# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "JLPT Mobile API",
        "description": "REST API for the JLPT mobile app. All authenticated endpoints expect a Sanctum bearer token issued by /auth/login or /auth/register.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://jlpt.senda.fit",
            "description": "Production"
        },
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local"
        }
    ],
    "paths": {
...
        "/api/v1/favorites": {
            "get": {
                "tags": [
                    "Favorites"
                ],
                "summary": "List the authenticated user's favorite vocab words (paginated)",
                "operationId": "8d2bd7cabe6dbeca723fdd10fb66150d",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/favorites/{vocabulary}": {
            "post": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Add a vocab word to favorites",
                "operationId": "83c8e4f26e91e47e80468e20a90c3faa",
                "parameters": [
                    {
                        "name": "vocabulary",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Added"
                    },
                    "404": {
                        "description": "Vocab not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Remove a vocab word from favorites",
                "operationId": "c1c7ea2765ba4686c45772d1374dcbb4",
                "parameters": [
                    {
                        "name": "vocabulary",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Removed"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Sanctum token",
                "scheme": "bearer"
            }
        }
    },
...
}